+++ tests-ui/tests/renderer/core/layout/utils/layoutMath.test.ts
@@ -3,24 +3,10 @@ import { describe, expect, it } from 'vitest'
 import {
   REROUTE_RADIUS,
   boundsIntersect,
-  makeLinkSegmentKey,
   pointInBounds
 } from '@/renderer/core/layout/utils/layoutMath'
 
 describe('layoutMath utils', () => {
-  describe('makeLinkSegmentKey', () => {
-    it('creates stable keys for null reroute', () => {
-      expect(makeLinkSegmentKey(10, null)).toBe('10:final')
-      expect(makeLinkSegmentKey(42, null)).toBe('42:final')
-    })
-
-    it('creates stable keys for numeric reroute ids', () => {
-      expect(makeLinkSegmentKey(10, 3)).toBe('10:3')
-      expect(makeLinkSegmentKey(42, 0)).toBe('42:0')
-      expect(makeLinkSegmentKey(42, 7)).toBe('42:7')
-    })
-  })
-
   describe('pointInBounds', () => {
     it('detects inclusion correctly', () => {
       const bounds = { x: 10, y: 10, width: 100, height: 50 }
